+++ prototypes/djerba_modular/plugin_hobbits/constants.py
@@ -1,5 +1,0 @@
-
-ROHAN = 'Esquire of Rohan'
-GONDOR = 'Fool of a Took'
-SHIRE = 'Mayor of the Shire'
-RING = 'Ringbearer'