+++ prototypes/djerba_modular/plugin_hobbits/ini.py
@@ -1,9 +1,0 @@
-"""INI parameter names"""
-
-FOO = 'foo'
-BAR = 'bar'
-BAZ = 'baz'
-FIZ = 'fiz'
-
-SCHEMA_REQUIRED = [FOO, BAR, BAZ]
-SCHEMA_OPTIONAL = [FIZ]