+++ prototypes/djerba_modular/plugin_wise/ini.py
@@ -1,8 +1,0 @@
-"""INI parameter names"""
-
-FOO = 'foo'
-BAR = 'bar'
-BAZ = 'baz'
-
-SCHEMA_REQUIRED = [FOO, BAR]
-SCHEMA_OPTIONAL = [BAZ]