+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## v1.1.0: 2023-11-23
+
+- removed `prototype/`
+- added RUO-report functionality
+- added HRD plugin to RUO report
+- split pWGS-sample plugin into three plugins
+
 ## v1.0.3: 2023-11-28
 
 - Allow tumour ID to be specified in the sample plugin